weather_project/requirements.txt
Django
requests
python-dotenv
gunicorn
whitenoise
psycopg2-binary